pyproject.toml
[project]
name = "vector-db-sizer"
version = "0.1.0"
description = "Analytical vector database storage estimator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7.0",
  "pyyaml>=6.0.0",
  "typer>=0.12.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
vector-db-sizer = "vector_db_sizer.cli:app"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
